reorg parser tests
