+++ dear_petition/portal/tests/parsers/conftest.py
@@ -1,0 +1,15 @@
+import pathlib
+
+from bs4 import BeautifulSoup
+import pytest
+
+
+@pytest.fixture(scope="module")
+def sample_record():
+    path = pathlib.Path(__file__).parent.parent / "data" / "record.html"
+    return path.read_text()
+
+
+@pytest.fixture(scope="module")
+def soup(sample_record):
+    return BeautifulSoup(sample_record, features="html.parser")
